Reverting back to default exports and moving to a single public file only
