+++ packages/workbox-background-sync/Queue.mjs
@@ -13,7 +13,7 @@
  limitations under the License.
 */
 
-import {WorkboxError} from 'workbox-core/_private/WorkboxError.mjs';
+import {WorkboxError} from 'workbox-core/_private.mjs';
 import {QueueStore} from './models/QueueStore.mjs';
 import StorableRequest from './models/StorableRequest.mjs';
 import {TAG_PREFIX, MAX_RETENTION_TIME} from './utils/constants.mjs';
