+++ packages/workbox-cache-expiration/CacheExpirationPlugin.mjs
@@ -11,8 +11,8 @@
  limitations under the License.
 */
 
-import {WorkboxError} from 'workbox-core/_private/WorkboxError.mjs';
-import {assert} from 'workbox-core/_private/assert.mjs';
+import {WorkboxError} from 'workbox-core/_private.mjs';
+import {assert} from 'workbox-core/_private.mjs';
 
 import CacheExpiration from './CacheExpiration.mjs';
 import './_version.mjs';
