+++ packages/workbox-core/_private/assert.mjs
@@ -14,7 +14,7 @@
   limitations under the License.
 */
 
-import {WorkboxError} from '../_private/WorkboxError.mjs';
+import WorkboxError from '../_private/WorkboxError.mjs';
 import '../_version.mjs';
 
 /*
@@ -87,4 +87,4 @@ const finalExports = process.env.NODE_ENV === 'production' ? null : {
   isType,
 };
 
-export {finalExports as assert};
+export default finalExports;
