+++ packages/workbox-core/browser.mjs
@@ -19,7 +19,6 @@ import LOG_LEVELS from './models/LogLevels.mjs';
 import * as _private from './_private.mjs';
 import './_version.mjs';
 
-
 const finalExports = Object.assign(defaultExport, {
   LOG_LEVELS,
   _private,
