+++ packages/workbox-routing/utils/normalizeHandler.mjs
@@ -13,7 +13,7 @@
  limitations under the License.
 */
 
-import {assert} from 'workbox-core/_private/assert.mjs';
+import {assert} from 'workbox-core/_private.mjs';
 import '../_version.mjs';
 
 /**
