+++ packages/workbox-runtime-caching/NetworkFirst.mjs
@@ -13,10 +13,10 @@
  limitations under the License.
 */
 
-import {getRuntimeName} from 'workbox-core/_private/cacheNames.mjs';
-import * as cacheWrapper from 'workbox-core/_private/cacheWrapper.mjs';
-import * as fetchWrapper from 'workbox-core/_private/fetchWrapper.mjs';
-import {assert} from 'workbox-core/_private/assert.mjs';
+import {cacheNames} from 'workbox-core/_private.mjs';
+import {cacheWrapper} from 'workbox-core/_private.mjs';
+import {fetchWrapper} from 'workbox-core/_private.mjs';
+import {assert} from 'workbox-core/_private.mjs';
 
 import cacheOkAndOpaquePlugin from './plugins/cacheOkAndOpaquePlugin.mjs';
 import './_version.mjs';
@@ -48,7 +48,7 @@ class NetworkFirst {
    * scenarios.
    */
   constructor(options = {}) {
-    this._cacheName = getRuntimeName(options.cacheName);
+    this._cacheName = cacheNames.getRuntimeName(options.cacheName);
 
     if (options.plugins) {
       let isUsingCacheWillUpdate =
