+++ packages/workbox-runtime-caching/StaleWhileRevalidate.mjs
@@ -13,9 +13,9 @@
  limitations under the License.
 */
 
-import {getRuntimeName} from 'workbox-core/_private/cacheNames.mjs';
-import * as cacheWrapper from 'workbox-core/_private/cacheWrapper.mjs';
-import * as fetchWrapper from 'workbox-core/_private/fetchWrapper.mjs';
+import {cacheNames} from 'workbox-core/_private.mjs';
+import {cacheWrapper} from 'workbox-core/_private.mjs';
+import {fetchWrapper} from 'workbox-core/_private.mjs';
 
 import cacheOkAndOpaquePlugin from './plugins/cacheOkAndOpaquePlugin.mjs';
 import './_version.mjs';
@@ -45,7 +45,7 @@ class StaleWhileRevalidate {
    * conjunction with this caching strategy.
    */
   constructor(options = {}) {
-    this._cacheName = getRuntimeName(options.cacheName);
+    this._cacheName = cacheNames.getRuntimeName(options.cacheName);
       this._plugins = options.plugins || [];
 
     if (options.plugins) {
