+++ test/workbox-precaching/node/test-precaching-module.mjs
@@ -2,7 +2,7 @@ import sinon from 'sinon';
 import clearRequire from 'clear-require';
 import {expect} from 'chai';
 
-import {assert} from '../../../packages/workbox-core/_private/assert.mjs';
+import assert from '../../../packages/workbox-core/_private/assert.mjs';
 
 describe(`[workbox-precaching] Module`, function() {
   let precachingModule;
