+++ test/workbox-precaching/node/utils/test-showWarningsIfNeeded.mjs
@@ -3,7 +3,7 @@ import {expect} from 'chai';
 
 import PrecacheEntry from '../../../../packages/workbox-precaching/models/PrecacheEntry.mjs';
 import showWarningsIfNeeded from '../../../../packages/workbox-precaching/utils/showWarningsIfNeeded.mjs';
-import {logger} from '../../../../packages/workbox-core/_private/logger.mjs';
+import logger from '../../../../packages/workbox-core/_private/logger.mjs';
 
 describe(`[workbox-precaching] showWarningsIfNeeded`, function() {
   let sandbox = sinon.sandbox.create();
